Changed : merge with Beliyaal branch Added : support for madVR (new Video Renderer)
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@1048 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/transform/mpadecfilter/dtsdec-0.0.1/libdts/dts_bitstream.c
@@ -43,7 +43,7 @@ void dts_bitstream_init (dts_state_t * state, uint8_t * buf, int word_mode,
     state->current_word = 0;
     state->word_mode = word_mode;
     state->bigendian_mode = bigendian_mode;
-    bitstream_get (state, align * 8);
+    bitstream_get (state, (uint32_t)(align * 8));
 }
 
 static inline void bitstream_fill_current (dts_state_t * state)
